Move 'user totp' to user_totp.go
